refactor containers, add row component
